requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
snowflake-connector-python[pandas]==3.12.4
streamlit==1.41.1
